remove redux
