+++ android/app/build.gradle
@@ -126,6 +126,7 @@ android {
 }
 
 dependencies {
+    compile project(':react-native-vector-icons')
     compile project(':react-native-fetch-blob')
     compile project(':react-native-image-picker')
     compile fileTree(dir: "libs", include: ["*.jar"])
