--- src/actions/index.js
@@ -1,1 +1,0 @@
-export * from './AddItemActions';
