--- src/actions/types.js
@@ -1,5 +1,0 @@
-// add item
-export const SEND_ITEM_FORM = 'send_item_form';
-export const FROM_TEXT_CHANGED = 'from_text_changed';
-export const ITEM_RESULTS = 'item_results';
-export const FETCH_ITEM_SUCCESS = 'fetch_item_success';
