+++ src/components/AddItemForm.js
@@ -6,17 +6,23 @@ import {
     Text,
     PixelRatio
 } from 'react-native';
-import { connect } from 'react-redux';
 import ImagePicker from 'react-native-image-picker';
-import {
-    sendItemForm,
-    isFromTextChanged,
-    itemResults
-} from '../actions';
+import RNFetchBlob from 'react-native-fetch-blob';
+import { Actions } from 'react-native-router-flux';
+import firebase from 'firebase';
 import { CardSection, Button, Input } from './common';
 
 
 class AddItemForm extends Component {
+
+    state = {
+        isFromText: '',
+        description: '',
+        responsePath: '',
+        avatarSource: null,
+        dbData: ''
+    }
+
     componentDidMount() {
         console.log('id: ', this.props.eventId);
     }
@@ -35,19 +41,64 @@ class AddItemForm extends Component {
 
 
         ImagePicker.showImagePicker(options, (response) => {
-            this.props.itemResults({ response });
+            console.log('Response = ', response);
+
+            if (response.didCancel) {
+                console.log('User cancelled image picker');
+            } else if (response.error) {
+                console.log('ImagePicker Error: ', response.error);
+            } else if (response.customButton) {
+                console.log('User tapped custom button: ', response.customButton);
+            } else {
+                const avatarSource = { uri: response.uri };
+                const responsePath = response.origURL;
+
+                this.setState({
+                    avatarSource,
+                    responsePath
+                });
+            }
         });
     }
 
     handleSendItemForm() {
-        const { isFrom, description, responsePath, eventId } = this.props;
+        const { description, responsePath, isFromText } = this.state;
+        const { eventId } = this.props;
 
-        this.props.sendItemForm({ isFrom, description, responsePath, eventId });
+        const testImageName = `image-from-react-native-${new Date()}.jpg`;
+        const { currentUser } = firebase.auth();
+        const path = `users/${currentUser.uid}/events/${eventId}/items/`;
+        const Blob = RNFetchBlob.polyfill.Blob;
+
+        // window.XMLHttpRequest = RNFetchBlob.polyfill.XMLHttpRequest;
+        // window.Blob = Blob;
+
+        Blob.build(RNFetchBlob.wrap(responsePath), { type: 'image/jpeg' })
+            .then((blob) => firebase.storage()
+                    .ref(path)
+                    .child(testImageName)
+                    .put(blob, { contentType: 'image/png' })
+            )
+            .then((snapshot) => {
+                // console.log(snapshot.downloadURL);
+                const itemURL = snapshot.downloadURL;
+                firebase.database().ref(path)
+                    .push({ name: isFromText, URL: itemURL })
+                    .then(() => {
+                        this.setState({
+                            isFromText: '',
+                            description: '',
+                            responsePath: '',
+                            avatarSource: null,
+                            dbData: ''
+                        });
+                        Actions.gifts({ eventId, type: 'reset' });
+                    });
+            });
     }
 
     render() {
         const { container, clothingItem, clothingItemContainer } = styles;
-        console.log('id: ', this.props.eventId);
         return (
             <View>
                 <CardSection>
@@ -57,11 +108,11 @@ class AddItemForm extends Component {
                             onPress={this.handleAddImageButton.bind(this)}
                         >
                             <View style={[container, clothingItem, clothingItemContainer]}>
-                                { this.props.avatarSource === null
+                                { this.state.avatarSource === null
                                     ? <Text>Select a Photo</Text>
                                     : <Image
                                         style={styles.clothingItem}
-                                        source={this.props.avatarSource}
+                                        source={this.state.avatarSource}
                                     />
                                 }
                             </View>
@@ -72,8 +123,8 @@ class AddItemForm extends Component {
                     <Input
                         placeholder="Bob"
                         label="From"
-                        value={this.props.isFrom}
-                        onChangeText={text => this.props.isFromTextChanged(text)}
+                        value={this.state.isFromText}
+                        onChangeText={isFromText => this.setState({ isFromText })}
                     />
                 </CardSection>
                 {/*<CardSection>
@@ -113,22 +164,4 @@ const styles = {
   }
 };
 
-const mapStateToProps = ({ addItem }) => {
-    const {
-        isFrom,
-        description,
-        responsePath,
-        avatarSource,
-    } = addItem;
-
-    return {
-        isFrom,
-        description,
-        responsePath,
-        avatarSource
-    };
-};
-
-export default connect(mapStateToProps, {
-    isFromTextChanged, sendItemForm, itemResults
-})(AddItemForm);
+export default AddItemForm;
