--- src/reducers/index.js
@@ -1,7 +1,0 @@
-import { combineReducers } from 'redux';
-import AddItemReducer from './AddItemReducer';
-
-
-export default combineReducers({
-    addItem: AddItemReducer
-});
